Feat: add cancle modal in postRegistraition Proccess
