+++ src/components/post/postRegister/PostRegisterForm.tsx
@@ -6,22 +6,16 @@ import CoffeeMenuSelection from '@/components/common/coffeeSelection/CoffeeMenuS
 import CoffeeOptionSelection from '@/components/common/coffeeSelection/CoffeeOptionSelection';
 import PostWriteSection from '@/components/post/postRegister/PostWriteSection';
 import Button from '@/components/common/Button';
-import ModalCTA from '@/components/common/ModalCTA';
 
 import { registPostState } from '@/atoms/atoms';
-import { BUTTON_TEXTS, MODAL_CTA_TEXTS } from '@/constants/common';
+import { BUTTON_TEXTS } from '@/constants/common';
 
 import { useUpadatePost } from '@/hooks/post/useUpadatePost';
 import { usePostMutation } from '@/hooks/post/usePostMutation';
-import { useVerifyModalCTA } from '@/hooks/useVerifyModalCTA';
-import { useSmartBack } from '@/hooks/post/useSmartBack';
-import { useResetRegistInfo } from '@/hooks/post/useResetRegistInfo';
 
 import { cx } from 'styled-system/css';
 import { styled } from 'styled-system/jsx';
 import { BottomBtnContainer, DefaultBtn, DisabledBtn } from '@/styles/styles';
-
-const { title, description } = MODAL_CTA_TEXTS.register;
 
 const PostRegisterForm = ({
   update,
@@ -40,17 +34,6 @@ const PostRegisterForm = ({
   const caffeineRegister = type === 'caffeine';
   useUpadatePost(update, postid);
 
-  // 등록 중단
-  const { isModal, setIsModal } = useVerifyModalCTA();
-  const { smartBack } = useSmartBack();
-  const { resetRegistInfo } = useResetRegistInfo();
-  const handleQuitbtn = () => {
-    isModal && setIsModal(!isModal);
-    resetRegistInfo();
-    smartBack();
-    return;
-  };
-
   // update && description update
   useEffect(() => {
     if (update && registInfo.description) {
@@ -59,26 +42,8 @@ const PostRegisterForm = ({
   }, [registInfo.description]);
 
   //포스트 등록 맟 수정 Hook
-  const {
-    mutate,
-    isPending,
-    userId,
-    registerProps,
-    cropperProps,
-    imageFile,
-    caffeine
-  } = usePostMutation(descriptions, update, caffeineRegister);
-
-  //유효성 검사
-  const hasContents = descriptions || imageFile || registInfo.photo;
-  const updateInvalid =
-    update &&
-    descriptions === registInfo?.description &&
-    (!imageFile || !registInfo?.photo);
-  const isInvalid =
-    (!update && !caffeine) ||
-    updateInvalid ||
-    (!caffeineRegister && !hasContents);
+  const { mutate, isPending, userId, registerProps, cropperProps, isInvalid } =
+    usePostMutation(descriptions, update, caffeineRegister);
 
   // 등록 및 수정 버튼
   const clickRegisterBtn = () => {
@@ -87,15 +52,6 @@ const PostRegisterForm = ({
 
   return (
     <>
-      {isModal && (
-        <ModalCTA
-          buttonText={[BUTTON_TEXTS.quit, BUTTON_TEXTS.continue]}
-          title={caffeineRegister ? title.caffeineIntake : title.post}
-          description={description}
-          type={BUTTON_TEXTS.type}
-          fn={handleQuitbtn}
-        />
-      )}
       <Container>
         <Update aria-disabled={update ? true : false}>
           <CoffeeMenuSelection />
@@ -115,7 +71,7 @@ const PostRegisterForm = ({
           text={!update ? BUTTON_TEXTS.registered : BUTTON_TEXTS.update}
           onClick={clickRegisterBtn}
           disabled={isInvalid}
-          className={cx(isInvalid && DisabledBtn, DefaultBtn)}
+          className={cx(!isInvalid ? DefaultBtn : DisabledBtn, DefaultBtn)}
         />
       </ButtonContainer>
     </>
